Add new component SelectAction for making select tag with options
